// packages/substrate/src/indexer/types.subscan.d.ts
export interface ApiScanExtrinsicResponse {
  code?: number
  data?: SubscanInternalModelExtrinsicDetail
  generated_at?: number
  message?: string
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelExtrinsicDetail {
  account_display?: SubscanInternalModelAccountDisplay
  account_id?: string
  additional_meta?: { [key: string]: any }
  block_hash?: string
  block_num?: number
  block_timestamp?: number
  call_module?: string
  call_module_function?: string
  error?: SubscanInternalModelExtrinsicError
  event?: SubscanInternalModelChainEventJson[]
  event_count?: number
  extrinsic_hash?: string
  extrinsic_index?: string
  fee?: number
  fee_used?: number
  finalized?: boolean
  lifetime?: SubscanInternalModelLifetime
  multisig?: SubscanInternalModelMultisigJson[]
  nonce?: number
  params?: SubscanInternalModelExtrinsicParam[]
  pending?: boolean
  proxy?: SubscanInternalModelExtrinsicProxyDetail
  signature?: string
  sub_calls?: SubscanInternalModelExtrinsicSubCallJson[]
  success?: boolean
  tip?: number
  transfer?: SubscanInternalModelTransferJson
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelAccountDisplay {
  account_index?: string
  /**
   * Current network account
   */
  address?: string
  display?: string
  evm_contract?: SubscanInternalModelEvmAccountDisplay
  identity?: boolean
  judgements?: SubscanInternalModelRegistrationJudgementJson[]
  merkle?: SubscanInternalModelMerkleTag
  /**
   * Parent account
   */
  parent?: Parent
  [property: string]: any
}
// ...
export interface SubscanInternalModelExtrinsicError {
  batch_index?: number
  doc?: string
  module?: string
  name?: string
  value?: string
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelChainEventJson {
  block_num?: number
  block_timestamp?: number
  event_id?: string
  event_idx?: number
  event_index?: string
  extrinsic_hash?: string
  extrinsic_idx?: number
  finalized?: boolean
  module_id?: string
  params?: string
  phase?: number
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelLifetime {
  birth?: number
  death?: number
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelMultisigJson {
  call_hash?: string
  call_module?: string
  call_module_function?: string
  multi_id?: string
  multisig_account_display?: SubscanInternalModelAccountDisplay
  multisig_status?: SubscanInternalModelMultiAction
  processing?: number
  threshold?: number
  [property: string]: any
}

/** @ignore */
export enum SubscanInternalModelMultiAction {
  Approval = 'Approval',
  Cancelled = 'Cancelled',
  Executed = 'Executed',
  Failed = 'Failed'
}

/** @ignore */
export interface SubscanInternalModelExtrinsicParam {
  name?: string
  type?: string
  type_name?: string
  value?: any
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelExtrinsicProxyDetail {
  account_display?: SubscanInternalModelAccountDisplay
  call_module?: string
  call_module_function?: string
  params?: SubscanInternalModelExtrinsicParam[]
  success?: boolean
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelExtrinsicSubCallJson {
  account?: SubscanInternalModelAccountDisplay
  exec_result?: SubscanLibsSubstrateMetadataModuleError
  exec_status?: string
  module?: string
  multisig?: SubscanInternalModelMultisigJson
  name?: string
  param?: any
  sub_calls?: SubscanInternalModelExtrinsicSubCallJson[]
  [property: string]: any
}

/** @ignore */
export interface SubscanLibsSubstrateMetadataModuleError {
  doc?: string[]
  module?: string
  name?: string
  value?: string
  [property: string]: any
}

/** @ignore */
export interface SubscanInternalModelTransferJson {
  amount?: number
  asset_symbol?: string
  from?: string
  hash?: string
  module?: string
  success?: boolean
  to?: string
  to_account_display?: SubscanInternalModelAccountDisplay
  [property: string]: any
}